Add --discard option for get_bulk command. to simulate /dev/null (JAVACLI-75)
