+++ ds3_java_cli/src/main/java/com/spectralogic/ds3cli/command/PerformanceCommand.java
@@ -7,6 +7,7 @@ import com.spectralogic.ds3cli.exceptions.CommandException;
 import com.spectralogic.ds3cli.models.PerformanceResult;
 import com.spectralogic.ds3cli.util.Ds3Provider;
 import com.spectralogic.ds3cli.util.FileUtils;
+import com.spectralogic.ds3cli.util.MemoryObjectChannelBuilder;
 import com.spectralogic.ds3client.Ds3Client;
 import com.spectralogic.ds3client.commands.DeleteBucketRequest;
 import com.spectralogic.ds3client.commands.DeleteObjectRequest;
@@ -139,21 +140,6 @@ public class PerformanceCommand extends CliCommand<PerformanceResult> {
         client.deleteBucket(new DeleteBucketRequest(bucketName));
     }
 
-    private class MemoryObjectChannelBuilder implements Ds3ClientHelpers.ObjectChannelBuilder {
-        private final int bufferSize;
-        private final long sizeOfFiles;
-
-        public MemoryObjectChannelBuilder(final int bufferSize, final long sizeOfFile) {
-            this.bufferSize = bufferSize;
-            this.sizeOfFiles = sizeOfFile;
-        }
-
-        @Override
-        public SeekableByteChannel buildChannel(final String key) throws IOException {
-            return new PerformanceSeekableByteChannel(this.bufferSize, this.sizeOfFiles);
-        }
-    }
-
     private class PerformanceListener implements DataTransferredListener, ObjectCompletedListener {
         private final long startTime;
         private final int totalNumberOfFiles;
@@ -210,63 +196,6 @@ public class PerformanceCommand extends CliCommand<PerformanceResult> {
         }
     }
 
-    private class PerformanceSeekableByteChannel implements SeekableByteChannel {
-        final private int bufferSize;
-        final private byte[] backingArray;
-        final private long limit;
-        private int position;
-        private boolean isOpen;
-
-        public PerformanceSeekableByteChannel(final int bufferSize, final long size) {
-            this.bufferSize = bufferSize;
-            final byte[] bytes = new byte[bufferSize];
-            for (int i = 0; i < bufferSize; i++) {
-                bytes[i] = '1';
-            }
-            backingArray = bytes;
-
-            this.position = 0;
-            this.limit = size * 1024L * 1024L;
-            this.isOpen = true;
-        }
-
-        public boolean isOpen() {
-            return this.isOpen;
-        }
-
-        public void close() throws IOException {
-            this.isOpen = false;
-        }
-
-        public int read(final ByteBuffer dst) throws IOException {
-            final int amountToRead = Math.min(dst.remaining(), this.bufferSize);
-            dst.put(this.backingArray, 0, amountToRead);
-            return amountToRead;
-        }
-
-        public int write(final ByteBuffer src) throws IOException {
-            final int amountToWrite = Math.min(src.remaining(), this.bufferSize);
-            src.get(this.backingArray, 0, amountToWrite);
-            return amountToWrite;
-        }
-
-        public long position() throws IOException {
-            return (long) this.position;
-        }
-
-        public SeekableByteChannel position(final long newPosition) throws IOException {
-            this.position = (int) newPosition;
-            return this;
-        }
-
-        public long size() throws IOException {
-            return this.limit;
-        }
-
-        public SeekableByteChannel truncate(final long size) throws IOException {
-            return this;
-        }
-    }
 }
